Rename packages for clarity
